Super Point CXX as a sub source folder
What:
1. add the cxx version of superpoint;
2. make the superpoint as a sub cmake folder;
3. add .gitignore for excluding the build, devel, logs, and etc.
Why:
The main reason to use superpoint is to make the code all cxx.
superpoint as a sub cmake folder is to make it an independent lib which can be used for other purpose.
Bugs:
The problem now is if linked the superpoint lib to the executable node, it will indicates the unreference to the ros::init and etc.